Sort timeline in ascending or descending order
Added support for specifying the direction in which timeline posts are
sorted.

The behaviour is controlled through the timeline_sorting configuration
option under [twtxt] in the config file. It can be ascending or
descending, and defaults to descending if no preference is specified.

The actual sorting is done after retrieving the entire list of tweets by
calling a helper function.

Signed-off-by: Weland Treebark <weland@blinkenshell.org>