Adding option to specify minimal interval between checks (closes #100)
The option is called 'timeline_update_interval', defaults to 10.
The behaviour is as follows: If cache is enabled and was updated
in the last 'timeline_update_interval' seconds, skip any network
requests and load the timline completely from the cache, otherwise
proceed with network requests and add new content to the cache, if
necessary. This only works if the cache is enabled.

An overwrite switch is also added, called: --force-update
When specified network requests will be made to check if
cache content is still up-to-date, but the cache will still
be used, when possible. Use --no-cache if you want to disable
the caching.